I am restructuring this class. parseInput was striping tags. After speaking with steven, needed to remove that function and put it in parseOutput Unfortunately, parseOutput, as a single function, was getting beefy.
So PHPWS_Text can now be used as an object to allow the dev to make several changes to the settings and better control display.
This new functionality is still in the testing stages.

Also added the parseTag and addTag functions. These allow you to create small bracketed tags that access functions within the cooresponding module. Thus you can insert elements of other modules into any place that holds content. Docs are needed for further explaination as well as usage of Clipboard to make this easier to use.