I rewrote a portion of this class. Hopefully it will increase flexibility and reduce complexitivity (that's a word, right?)
Instead of addRowTag and setMethod there is just one function: addRowTags. You just send it the name of a method in your target object and it expects an associative array of tags per row. This tags can be in addition to or instead of the tags pulled from the table.