pear's template class uses loadTemplateFile while I was using loadTemplatefile. Although it worked fine, I figured they should match.
Added an error check to the process() function. If you used repeating rows and the repeating row tags were not in your template, the template class would not work. I now check to make sure the repeat tag is contained in the template objects _block variable. Unfortunately this is a "hidden" variable