Update User.php
This isn't 100% right, but allows the month/week/day buttons to appear in those views. However, the month grid button always shows you the month you are on whereas the month button takes you to the list of the current month. 

I'll be sending updated versions of the templates as well. I'm hoping you can make the month grid/month button toggle between list and grid for the month that is displayed. I'm also having a problem getting the buttons to display on the day if there are events. if there are no events, month and week, and the arrow buttons display.