Added new parameter to Layout functions/methods
javascript, Layout::getJavascript, and Layout::loadJavascriptFile function now have a wrap_content parameter. Defaulted to false, but if true, the head.js will get wrapped with <script> tags.
This allows us to write straight javascript files without <script> tags within them. A .JS file with tags is not semantically correct, it identified as an error in code editors, and more difficult to edit as a result.
As false is the default condition, this will not effect old code.

Other edits are style related, less long line breaking.