getDataType and alter methods were worked on.
In Postgresql's getDataType, the process for finding default is flawed in the case of indexed integers. The ability to get specific information on columns is dirty so you have to parse the information returned. So, in the case of primary key integers, we don't set the default because the default is always the returned "next key."

Removed the setIsPrimaryKey and setIsUnique. Neither flag is in Datatype anymore. Primary key I understand as you may have multiple primary key columns. Unique removal I am not so sure of. In any case, they causes errors and were removed.

A little tinkering in the alterDefaultStatus method. If the new datatype has a unset default (null) then we don't run a default update.