Fixed bug with Global/Database classes and primary key sequences.
The Global database classes expected all tables to have an auto_increment (mysql) or a serial (pgsql) column for their primary keys. In the case of old phpwebsite tables, this wasn't the case. Old phpwebsite tables depended on MDB2's sequence handling. If the insert method was called in the Global database classes, you would either get a 0 for an id or an error for not including the id column.

The update adds in new methods to all inserts to work with old phpwebsite tables.
Insert will now check usePearSequence has been set on old tables. If not, then Global will not allow inserts until it is. Note, this check will be ignored on tables created with Global as they use proper serial columns. If usePearSequence(TRUE) is called, then the MDB2 sequence table will be accessed and updated properly and the insert query will be updated properly. If it is set to FALSE, then it will plug ahead with, most likely, a bad insert.

Also added:
- DB::getPDO loads the PDO if it is missing.
- the Postgresql hasSequenceTable method would unload the PDO static object on return. Thorough testing could not reveal why this happened but it could be narrowed down to the creation of a new DB object. Fixed the bug by running the query through the PDO directly.