Added longtext and mediumtext to postgresql.
This is unorthodox, but there may be a need to create a medium or long text column in mysql. PSQL doens't have those, it just has TEXT (which is better).
In order to allow the creation of a column bigger than mysql's TEXT, the postgresql additions were made to prevent an error.