Rewrote Bool to work like php boolean logic
PHP has different rules for detecting and transforming boolean values. The
Bool Variable was rewritten to allow some of PHP logic but not other.

For Bool, the string 'false' is boolean and false. 'foobar' is not boolean.
PHP would see 'false' and boolean and false but 'foobar' as not boolean and true
Bool will on accept any strings besides yes, no, true, and false.