Added loadPostByType
Pulls POST varaibles from the $request and tries to set
the variable in the Resource object.
An ignore array parameter can be passed to forgive missing post variables.
If post variables not in the resource will automatically be ignored.
Exceptions are thrown if the post variable doesn't match the Variable type
or if the resource Variable is of an unknown type.