Changed option set to default 0
How did this ever work? The database table is an integer. Always has been. But
the element was set to an array as a default and then it tried to update as
a serialize array (a string). I am putting the notes here unless there is
deeper problem I am not aware of.